[swisstopo] Organize
