+++ iOS/Multiplatform Redux Sample/Multiplatform Redux Sample/Main/CollectionPointInfoView.swift
@@ -8,91 +8,10 @@
 import ReduxSampleShared
 import UIKit
 
-class IconStack: UIView {
-    let titleLabel = UILabel.paragraph2()
-    let icons = UIStackView.autoLayout(axis: .horizontal)
-    var iconBackgroundColor = UIColor.clear
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubview(titleLabel)
-        addSubview(icons)
-        icons.alignment = .center
-        icons.distribution = .fill
-        icons.spacing = kUnit1
-
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: topAnchor),
-            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor),
-            titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor),
-            icons.leadingAnchor.constraint(equalTo: leadingAnchor),
-            icons.trailingAnchor.constraint(lessThanOrEqualTo: trailingAnchor),
-            icons.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 8),
-            icons.bottomAnchor.constraint(equalTo: bottomAnchor),
-        ])
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func setIcons(from imageNames: [String]) {
-        icons.removeAllArrangedSubviews()
-        imageNames.compactMap { name in
-            let imageView = RoundColoredImage(withSize: kUnit4, iconSize: kUnit3, backgroundColor: iconBackgroundColor)
-            imageView.setImage(name: name)
-            return imageView
-        }.forEach { icon in
-            icons.addArrangedSubview(icon)
-        }
-    }
-}
-
-class HorizontalDoublekView<T>: UIView where T: UIView {
-    let leading = T.autoLayout()
-    let trailing = T.autoLayout()
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        addSubview(leading)
-        addSubview(trailing)
-
-        NSLayoutConstraint.activate([
-            leading.leadingAnchor.constraint(equalTo: leadingAnchor),
-            leading.topAnchor.constraint(equalTo: topAnchor),
-            leading.bottomAnchor.constraint(equalTo: bottomAnchor),
-            leading.trailingAnchor.constraint(equalTo: centerXAnchor),
-            trailing.leadingAnchor.constraint(equalTo: centerXAnchor),
-            trailing.topAnchor.constraint(equalTo: topAnchor),
-            trailing.bottomAnchor.constraint(equalTo: bottomAnchor),
-            trailing.trailingAnchor.constraint(equalTo: trailingAnchor),
-        ])
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension UIButton {
-    class func createLink() -> UIButton {
-        let button = UIButton(type: .custom)
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.titleLabel?.font = .link()
-        button.contentHorizontalAlignment = .left
-        button.setTitleColor(.accentDarkAccent, for: .normal)
-        button.setTitleColor(.accentAccent, for: .highlighted)
-
-        return button
-    }
-}
-
 class CollectionPointInfoView: UIView {
     let stack = UIStackView.autoLayout(axis: .vertical)
     let titleLabel = UILabel.h4()
-    let iconStacks = HorizontalDoublekView<IconStack>.autoLayout()
+    let iconStacks = HorizontalDoublekView<IconStackView>.autoLayout()
     let addressLabel = UILabel.paragraph2()
     let mapLink = UIButton.createLink()
     let closeControl = UIButton(type: .custom)
@@ -124,14 +43,13 @@ class CollectionPointInfoView: UIView {
     }
 
     func render(_ viewState: CollectionPointViewState) {
-        kermit().d(viewState)
         titleLabel.text = viewState.title
         addressLabel.text = viewState.address
         mapLink.setTitle(viewState.navigationLink.text.localized, for: .normal)
 
         let typeStack = iconStacks.leading
         typeStack.iconBackgroundColor = .primaryLight
-        typeStack.titleLabel.text = viewState.collectionPointTypeTitle(localize: Localiser())
+        typeStack.titleLabel.text = viewState.collectionPointTypeTitle(localize: Localizer())
         typeStack.setIcons(from: viewState.collectionPointTypes.map { $0.icon })
 
         let wheelchairStack = iconStacks.trailing
@@ -139,11 +57,5 @@ class CollectionPointInfoView: UIView {
         wheelchairStack.titleLabel.text = viewState.wheelChairAccessibleTitle.localized
         wheelchairStack.setIcons(from: [viewState.wheelChairAccessibleIcon])
         wheelchairStack.alpha = viewState.wheelChairAccessible ? 1 : 0.3
-    }
-}
-
-class Localiser: Localize {
-    func localize(string: String) -> String {
-        string.localized
     }
 }
